Update Aid Calculator.html
actually change those controls to a slider, from 0 to 1 million with 10000 as a unit. link the slider value with value in the input box

pre-populate the family income and asset input text boxes with the same value as in the sliders, 400000 for income, 750000 for asset. also increase the asset upper limit to 4 million.